fix: refactor lazy imports to use common.py correctly.
