+++ honcho-cli/src/honcho_cli/common.py
@@ -1,4 +1,4 @@
-"""Shared callback and command-level flag helpers.
+"""Shared runtime state, client factory, and command-level flag helpers.
 
 Flags like --json, -w, -p, -s work in TWO positions:
   1. Group-level (before subcommand):  honcho workspace --json list
@@ -16,17 +16,69 @@ import typer
 from honcho_cli.output import set_json_mode
 
 
+# Global overrides from flags (commands read these)
+_global_overrides: dict[str, str | None] = {
+    "workspace": None,
+    "peer": None,
+    "session": None,
+}
+
+
+def get_resolved_config():
+    """Get config with global flag overrides applied.
+
+    Overrides flow through ``validate_resource_id`` so that a malformed
+    ``-w``/``-p``/``-s`` value fails fast with a structured error rather than
+    reaching the API and surfacing as an opaque ``UNKNOWN_ERROR``.
+    """
+    from honcho_cli.config import CLIConfig
+    from honcho_cli.validation import validate_resource_id
+
+    config = CLIConfig.load()
+
+    if _global_overrides["workspace"]:
+        config.workspace_id = validate_resource_id(_global_overrides["workspace"], "workspace")
+    if _global_overrides["peer"]:
+        config.peer_id = validate_resource_id(_global_overrides["peer"], "peer")
+    if _global_overrides["session"]:
+        config.session_id = validate_resource_id(_global_overrides["session"], "session")
+
+    return config
+
+
+def get_client(*, require_workspace: bool = True):
+    """Create a Honcho client from resolved config.
+
+    By default, refuses to build a client when no workspace is scoped — the
+    SDK's get-or-create semantics would otherwise silently operate on an empty
+    workspace. Commands that legitimately run without a workspace (e.g.
+    ``workspace list``) pass ``require_workspace=False``.
+    """
+    from honcho import Honcho
+
+    from honcho_cli.config import get_client_kwargs
+    from honcho_cli.output import print_error
+
+    config = get_resolved_config()
+    if require_workspace and not config.workspace_id:
+        print_error(
+            "NO_WORKSPACE",
+            "No workspace scoped. Pass --workspace/-w or set HONCHO_WORKSPACE_ID.",
+        )
+        raise typer.Exit(1)
+    return Honcho(**get_client_kwargs(config)), config
+
+
 def handle_cmd_flags(
     json_output: bool = False,
     workspace: str | None = None,
     peer: str | None = None,
     session: str | None = None,
+    **_kwargs,
 ) -> None:
     """Apply command-level flags. Idempotent if already set by group callback."""
     if json_output:
         set_json_mode(True)
-
-    from honcho_cli.main import _global_overrides
 
     if workspace:
         _global_overrides["workspace"] = workspace
@@ -50,8 +102,6 @@ def add_common_options(app: typer.Typer) -> None:
         if json_output:
             set_json_mode(True)
 
-        from honcho_cli.main import _global_overrides
-
         if workspace:
             _global_overrides["workspace"] = workspace
         if peer:
@@ -60,4 +110,4 @@ def add_common_options(app: typer.Typer) -> None:
             _global_overrides["session"] = session
 
         if ctx.invoked_subcommand is None:
-            ctx.get_help()
+            typer.echo(ctx.get_help())
