+++ honcho-cli/src/honcho_cli/main.py
@@ -11,6 +11,7 @@ import typer
 
 from honcho_cli import __version__
 from honcho_cli.branding import BANNER, BRAND
+from honcho_cli.common import _global_overrides
 from honcho_cli.output import set_json_mode
 
 app = typer.Typer(
@@ -73,61 +74,6 @@ def main(
             console.print(f"  [dim]v{__version__}[/dim]\n")
         console.print(ctx.get_help())
         raise typer.Exit()
-
-
-# Global overrides from flags (commands read these)
-_global_overrides: dict[str, str | None] = {
-    "workspace": None,
-    "peer": None,
-    "session": None,
-}
-
-
-def get_resolved_config():
-    """Get config with global flag overrides applied.
-
-    Overrides flow through ``validate_resource_id`` so that a malformed
-    ``-w``/``-p``/``-s`` value fails fast with a structured error rather than
-    reaching the API and surfacing as an opaque ``UNKNOWN_ERROR``.
-    """
-    from honcho_cli.config import CLIConfig
-    from honcho_cli.validation import validate_resource_id
-
-    config = CLIConfig.load()
-
-    if _global_overrides["workspace"]:
-        config.workspace_id = validate_resource_id(_global_overrides["workspace"], "workspace")
-    if _global_overrides["peer"]:
-        config.peer_id = validate_resource_id(_global_overrides["peer"], "peer")
-    if _global_overrides["session"]:
-        config.session_id = validate_resource_id(_global_overrides["session"], "session")
-
-    return config
-
-
-def get_client(*, require_workspace: bool = True):
-    """Create a Honcho client from resolved config.
-
-    By default, refuses to build a client when no workspace is scoped — the
-    SDK's get-or-create semantics would otherwise silently operate on an empty
-    workspace. Commands that legitimately run without a workspace (e.g.
-    ``workspace list``) pass ``require_workspace=False``.
-    """
-    import typer
-
-    from honcho import Honcho
-
-    from honcho_cli.config import get_client_kwargs
-    from honcho_cli.output import print_error
-
-    config = get_resolved_config()
-    if require_workspace and not config.workspace_id:
-        print_error(
-            "NO_WORKSPACE",
-            "No workspace scoped. Pass --workspace/-w or set HONCHO_WORKSPACE_ID.",
-        )
-        raise typer.Exit(1)
-    return Honcho(**get_client_kwargs(config)), config
 
 
 # Register top-level commands
